Added a blur effect on the typing box text if the content is not focused.
